// docs/review_bundle/types.ts
export interface Product {
    product_id: string
    sku_code: string
    product_name: string
    category: string
    status: string
    launch_month: string
    sales_channel: string
    price: string
    [key: string]: any
}

export interface Task {
    product_task_id: string
    product_id: string
    task_name: string
    status: string
    due_date: string
    owner_email: string
    [key: string]: any
}

export interface Attachment {
    attachment_id: string
    product_id: string
    type: string
    drive_url: string
    created_at: string
    [key: string]: any
}

export interface SalesItem {
    order_id: string
    order_date: string
    sku: string
    quantity: string
    total_amount: string
    [key: string]: any
}